medmapp/apps/blog/models.py
from django.db import models
from users.models import User

# Create your models here.
from ckeditor_uploader.fields import RichTextUploadingField

class Post(models.Model):
    title = models.CharField(max_length=255)
    body = RichTextUploadingField(blank=True)  # supports image upload
    author = models.ForeignKey(User, on_delete=models.CASCADE)
    created_at = models.DateTimeField(auto_now_add=True)

    def __str__(self):
        return self.title
