medmapp/requirements.txt
cloudinary==1.44.1
Django==5.2.7
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
pillow==12.0.0
python-decouple==3.8
requests==2.32.5
urllib3==2.5.0
